api/fast-api-server/requirements.txt
fastapi
pydantic
httpx
python-dotenv